msp430: Add digital I/O ports
This adds support for the MSP430 Digital I/O ports, i.e. the GPIOs. On
most MCUs, only the first two ports support interrupts. Ports 7, 8, 9,
and 10 may be combined in pairs, such that registers for ports 7/9 use
even addresses and registers for ports 8/10 use odd addresses. This
arrangement is facilitated by the port16 device.

One unusual feature of the I/O ports is that not all registers are
modified on reset. They remain unchanged by a power-up clear (warm
reset) and are set to random values on cold reset to emulate their
indeterminate values on real hardware.

Over several generations, additional registers have been added to the
I/O ports. As there was no remaining address space in the original
locations, these registers are discontiguously located in other parts of
the address space. This is implemented with two additional address
spaces.

One of these registers is the REN or reistsor enable register. This
controls an internal pull-up/-down resistor. Any negative value on an
input irq will be considered a high impedance state, and if the resistor
is enabled it will set the input value as read by the CPU.

The "SEL" registers implement pin-muxing capabilities. They are not
currently implemented.

Signed-off-by: Sean Anderson <seanga2@gmail.com>