msp430: Add virt machine
This adds a "virt" machine, which represents an MCU soldered down on a
bare board with an optional crystal. This is a common form factor for
manu MSP430 breakout boards. In particular, most of Olimex's "header"
boards can be modeled this way. Some of TI's "Launchpad" boards are also
quite simple, although they typically include LEDs and (unmodeled)
buttons.

Signed-off-by: Sean Anderson <seanga2@gmail.com>