msp430: Add MPY (multiplier)
This adds support for the MSP430 hardware multiplier peripheral, as
implemented in the MSP430x2xx and MSP430x4xx families. The MSP430
instruction set has no multiplication instructions. Instead, there is a
device present on some microcontrollers which can be used to perform
multiplication.

The original device (MPY) could perform 8x8 or 16x16 multiplication (and
accumulation). This is done by writing the operands and then reading
back the results:

    MOV   #0x1234, &MPY ; Load first operand
    MOV   #0x5678, &OP2 ; Load second operand;
    MOV   &RESLO,  R6   ; R6 = 0x0060
    MOV   &RESHI,  R7   ; R7 = 0x0626

Unfortunately, interrupts have to be disabled while using the multiplier
because the type of multiplication is determined based on the register
written to, and this state is not exposed anywhere else. If an interrupt
were to occur in the above example between writing MPY and OP2, there
would be no way to use the multiplier without potentially corrupting the
results. Additionally, byte writes to MPYS and MACS (or OP2 after a
write to those registers) are not automatically sign-extended.

TI sought to address these issues with MPY32, which (in addition to
adding support for 32x32 multiplication and fractional/saturating modes)
exposes the internal multiplier state in the MPY32CTL0 register. This
would allow use of the multiplier with interrupts enabled if the
interrupt routine saved the MPY32 registers before using the multiplier
and restored them after. Unfortunately, they forgot a bit
(expecting_op2h), so this cannot be done.

The fractional and saturating modes are a bit unusual. It is not quite
clear from TI's literature (e.g. slau056l) how the multiplication
results are stored. For example, the fractional mode is documented as:

> The actual content of the result register(s) is not modified when
> MPYFRAC = 1. When the result is accessed using software, the value is
> left−shifted 1 bit resulting in the final Q formatted result. This allows
> user software to switch between reading both the shifted (fractional) and
> the un-shifted result

However, later on Figure 9−5 ("Multiplication Flow Chart") shows this
left-shifting being applied to the result as part of the multiplication
process. I have chosen the interpretation as noted in the above quote, but
experimentation is needed to determine the actual behavior.

Additrionally, experimentation is also needed to determine the correct
behavior of byte writes, especially "unaligned" byte writes. Although this
is a 16-bit peripherals, 8-bit writes are allowed to some registers (to
facilitate 8-bit multiplication). However, it is unclear which registers
(if any) require 16-bit writes. Examples from TI only show byte accesses to
operand registers, and only to even addresses. However, some user guides
(e.g. slau445i) document byte registers at odd offsets. I have chosen to
implement byte reads/writes for all registers, but this could be
incorrect.

The multiplier takes several cycles to perform multiplication. Although
this is usually negligable for 16x16 multiplications (where the time taken
to execute the instruction accessing the result registers is enough to add
the necessary delay), it must be accounted for when performing 32x32
multiplications (which are calculated iteratively). I did not model this at
all, since QEMU does not have precise enough timing at this scale.

Signed-off-by: Sean Anderson <seanga2@gmail.com>