display: Add support for GPIO-controlled LCDs
This adds support for LCDs with direct external control of segments.
This is intended to model LCDs with common and segment lines, although
no specific topology is enforced.

The size and shape of the segments are configured with an "lcdmap": an
indexed-color PNG. The index of the color of each pixel determines the
segment it is part of. This limits us to LCDs with at most 255
segments (with one color reserved for the background). It also means
that we cannot support anti-aliasing in the lcdmap (since aliased pixels
would be different colors).

The first limitation could be addressed by using regular PNGs, and
storing the mapping from colors to segments as class data. This would
require a conversion from full color to indexed color when realizing the
device.

The second limitation could be addressed by postprocessing the output to
apply anti-aliasing. However, I do not find the segments to be
particularly distracting on the example board.

To find the lcdmap, we add a new file type to datadir. An initial map is
supplied for the SoftBaugh SBLCDA2 LCD, present on their ES449 board.

At the moment, the while display is invalidated whenever any segment
changes. As an optimization, it would be possible to record the
bounding-boxes of each segment and only redraw those boxes when a
segment changed.

We unfortunately can't depend on CONFIG_PNG, since it's not a real
Kconfig. Instead, we allow the file to be built and fail at runtime.
This is less than ideal.

I'm not sure how to support VM migration, as the lcdmap is of variable
size in two dimensions (so things like VMSTATE_UINT32_2DARRAY_V won't
work). Suggestions welcome.

Signed-off-by: Sean Anderson <seanga2@gmail.com>