msp430: Add LCD (controller)
This adds support for the MSP430 "LCD" controller found on MSP430x4xx
microcontrollers. It supports up to 4 common lines and 40 segment lines,
for a total of 160 possible segments. Enabled common lines have an AC
waveform applied, with each line's waveform being orthogonal to the
others. Segments are turned on by outputting an AC waveform either
reinforcing or canceling-out the effect of the corresponding common
line.

Rather than model the common/segment lines individually, I have chosen
to just model the resulting segment values. This effectively turns the
whole device into a GPO controller. The frame rate is not modeled, and
the clock just needs to be enabled for the device to function.

Although the contrast of the display can be adjusted with external
resistors (or registers settings in the case of the similar LCD_A
device), this is not modeled.

Signed-off-by: Sean Anderson <seanga2@gmail.com>